requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-timezone-field==4.1.1
djangorestframework==3.11.0
flake8==3.8.4
flower==0.9.7
humanize==3.2.0
idna==2.10
kombu==4.6.11
mccabe==0.6.1
Pillow==8.1.0
prometheus-client==0.8.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.2
vine==1.3.0
